# Vectorize Convert.ToBase64CharArray and TryToBase64Chars