https://github.com/dotnet/runtime/pull/71795 vectorized Convert.ToBase64String for larger inputs by using Base64.EncodeToUTF8 and then encoding the result UTF8 bytes into a UTF16 string.  It did not touch Convert.ToBase64CharArray nor Convert.TryToBase64Chars, however.  The ToBase64String change makes use of a temporary array rented from the array pool, and the expectation is it'll rarely allocate, but if it does, it's part of a method that's already allocating the resulting string and so it's presumed to not be too impactful.  ToBase64CharArray and TryToBase64Chars, however, are intended to be entirely non-allocating, and so even renting from the array pool could be problematic if it fails to find a buffer in the pool.

This PR changes the non-allocating variants to use Base64.EncodeToUtf8 as well.  But instead of renting a temporary buffer, it banks on the knowledge that the encoded Base64 bytes are 1/2 the length of the resulting chars, since the bytes are all guaranteed to be ASCII.  Thus, it can treat the destination char buffer as scratch space for the encoded UTF8 bytes, and then widen in-place.  This obviates the need for a separate temporary buffer, making it appropriate for the non-allocating versions.  And once we had the helper for those, we can use that same helper to replace the code added to ToBase64String, making it non-allocating as well (beyond of course the result string it has to allocate by its nature), and thus making it more predictable.

Overall, this fixes the possible additional allocation in ToBase64String as well as the performance inversion that the allocating ToBase64String could have been significantly faster (due to vectorization) than the ToBase64CharArray and ToBase64Chars methods intended to be the faster versions.

```C#
[Params(16, 64, 256, 1024)]
public int Length { get; set; }

private byte[] _data;
private char[] _scratch;

[GlobalSetup]
public void Setup()
{
    _data = new byte[Length];
    _scratch = new char[Length * 4];
    var r = new Random(42);
    r.NextBytes(_data);
}

[Benchmark]
public string ToBase64String() => Convert.ToBase64String(_data);

[Benchmark]
public void ToBase64CharArray() => Convert.ToBase64CharArray(_data, 0, _data.Length, _scratch, 0);

[Benchmark]
public void ToBase64Chars() => Convert.TryToBase64Chars(_data, _scratch, out _);
```

|            Method |         Toolchain | Length |        Mean | Ratio | Allocated |
|------------------ |------------------ |------- |------------:|------:|----------:|
|    ToBase64String | \main\corerun.exe |     16 |    34.86 ns |  1.00 |      72 B |
|    ToBase64String |   \pr\corerun.exe |     16 |    34.82 ns |  1.00 |      72 B |
|                   |                   |        |             |       |           |
| ToBase64CharArray | \main\corerun.exe |     16 |    26.08 ns |  1.00 |         - |
| ToBase64CharArray |   \pr\corerun.exe |     16 |    27.12 ns |  1.04 |         - |
|                   |                   |        |             |       |           |
|     ToBase64Chars | \main\corerun.exe |     16 |    25.67 ns |  1.00 |         - |
|     ToBase64Chars |   \pr\corerun.exe |     16 |    26.55 ns |  1.03 |         - |
|                   |                   |        |             |       |           |
|    ToBase64String | \main\corerun.exe |     64 |    50.12 ns |  1.00 |     200 B |
|    ToBase64String |   \pr\corerun.exe |     64 |    49.22 ns |  0.98 |     200 B |
|                   |                   |        |             |       |           |
| ToBase64CharArray | \main\corerun.exe |     64 |    79.72 ns |  1.00 |         - |
| ToBase64CharArray |   \pr\corerun.exe |     64 |    31.39 ns |  0.39 |         - |
|                   |                   |        |             |       |           |
|     ToBase64Chars | \main\corerun.exe |     64 |    78.80 ns |  1.00 |         - |
|     ToBase64Chars |   \pr\corerun.exe |     64 |    31.49 ns |  0.40 |         - |
|                   |                   |        |             |       |           |
|    ToBase64String | \main\corerun.exe |    256 |   137.63 ns |  1.00 |     712 B |
|    ToBase64String |   \pr\corerun.exe |    256 |   108.71 ns |  0.79 |     712 B |
|                   |                   |        |             |       |           |
| ToBase64CharArray | \main\corerun.exe |    256 |   300.65 ns |  1.00 |         - |
| ToBase64CharArray |   \pr\corerun.exe |    256 |    47.43 ns |  0.16 |         - |
|                   |                   |        |             |       |           |
|     ToBase64Chars | \main\corerun.exe |    256 |   299.34 ns |  1.00 |         - |
|     ToBase64Chars |   \pr\corerun.exe |    256 |    46.80 ns |  0.16 |         - |
|                   |                   |        |             |       |           |
|    ToBase64String | \main\corerun.exe |   1024 |   392.78 ns |  1.00 |    2760 B |
|    ToBase64String |   \pr\corerun.exe |   1024 |   346.42 ns |  0.88 |    2760 B |
|                   |                   |        |             |       |           |
| ToBase64CharArray | \main\corerun.exe |   1024 | 1,174.50 ns |  1.00 |         - |
| ToBase64CharArray |   \pr\corerun.exe |   1024 |   116.84 ns |  0.10 |         - |
|                   |                   |        |             |       |           |
|     ToBase64Chars | \main\corerun.exe |   1024 | 1,162.84 ns |  1.00 |         - |
|     ToBase64Chars |   \pr\corerun.exe |   1024 |   116.44 ns |  0.10 |         - |